% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_probs}
\alias{grid_probs}
\title{Function to compute predicted quantiles}
\usage{
grid_probs(X, q, delta, median_loc)
}
\arguments{
\item{X}{Matrix of class predictions, where each column gives the predictions
for a given quantile in q.}

\item{q}{The quantiles for which the columns of X are predictions.}

\item{delta}{The number of quantiles used.}

\item{median_loc}{Location of median quantile (0-based indexing).}
}
\description{
Find predicted quantiles given classification results at different quantiles.
}
