% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jous_utility.R
\name{index_over}
\alias{index_over}
\title{Return indices to be used for jittered data in oversampling}
\usage{
index_over(ix_pos, ix_neg, q)
}
\arguments{
\item{ix_pos}{Indices for positive examples in data.}

\item{ix_neg}{Indices for negative examples in data.}

\item{q}{Quantiles for which to construct tilted datasets.}
}
\value{
returns a list, each of element of which gives indices to be used on
        a particular cut (note: will be of length delta - 1)
}
\description{
Return indices to be used for jittered data in oversampling
}
