% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOUSBoost.R
\docType{package}
\name{JOUSBoost}
\alias{JOUSBoost}
\alias{JOUSBoost-package}
\title{JOUSBoost: A package for probability estimation}
\description{
JOUSBoost implements under/oversampling with jittering for probability estimation.
 Its intent is to be used to improve probability estimates that come from
boosting algorithms (such as AdaBoost), but is modular enough to be used with
virtually any classification algorithm from machine learning.
}
\details{
For more theoretical background, consult Mease (2007).
}
\references{
Mease, D., Wyner, A. and Buja, A. (2007). Costweighted
boosting with jittering and over/under-sampling:
JOUS-boost. J. Machine Learning Research 8 409-439.
}
