% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost.R
\name{print.AdaBoost}
\alias{print.AdaBoost}
\title{Print a summary of adaboost fit.}
\usage{
\method{print}{AdaBoost}(x, ...)
}
\arguments{
\item{x}{An AdaBoost object fit using the \code{adaboost} function.}

\item{...}{\dots}
}
\value{
Printed summary of the fit, including information about the tree
        depth and number of boosting rounds used.
}
\description{
Print a summary of adaboost fit.
}
