% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jous.R
\name{print.JOUS}
\alias{print.JOUS}
\title{Print a summary of \code{jous} fit.}
\usage{
\method{print}{JOUS}(x, ...)
}
\arguments{
\item{x}{A JOUS object fit using the \code{jous} function.}

\item{...}{\dots}
}
\value{
Printed summary of the fit
}
\description{
Print a summary of \code{jous} fit.
}
