% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOUSBoost.R
\docType{data}
\name{sonar}
\alias{sonar}
\title{Dataset of sonar measurements of rocks and mines}
\format{A data frame with 208 observations on 61 variables.  The variables
        V1-V60 represent the energy within a certain frequency band, and
        are to be used as predictors.  The variable y is a class label, 1
        for 'rock' and -1 for 'mine'.}
\source{
\url{http://archive.ics.uci.edu/ml/}
}
\usage{
data(sonar)
}
\description{
A dataset containing sonar measurements used to discriminate rocks from mines.
}
\references{
Gorman, R. P., and Sejnowski, T. J. (1988). "Analysis of Hidden
 Units in a Layered Network
 Trained to Classify Sonar Targets" in Neural Networks, Vol. 1, pp. 75-89.
}
\keyword{datasets}
