% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jous_utility.R
\name{index_under}
\alias{index_under}
\title{Return indices to be used in original data for undersampling}
\usage{
index_under(ix_pos, ix_neg, q, delta)
}
\arguments{
\item{ix_pos}{Indices for positive examples in data.}

\item{ix_neg}{Indices for negative examples in data.}

\item{q}{Quantiles for which to construct tilted datasets.}

\item{delta}{Number of quantiles.}
}
\value{
returns a list, each of element of which gives indices to be used on
        a particular cut (note: will be of length delta - 1)
}
\description{
(note: sampling is done without replacement)
}
