% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaBoost.R
\name{print.AdaBoost}
\alias{print.AdaBoost}
\title{Print a summary of adaBoost fit.}
\usage{
\method{print}{AdaBoost}(x, ...)
}
\arguments{
\item{x}{An AdaBoost object fit using the \code{adaBoost} function.}

\item{...}{\dots}
}
\value{
Printed summary of the fit, including information about the tree
        depth and number of boosting rounds used.
}
\description{
Print a summary of adaBoost fit.
}

