% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jous.R
\name{print.jous}
\alias{print.jous}
\title{Print a summary of \code{jous} fit.}
\usage{
\method{print}{jous}(x, ...)
}
\arguments{
\item{x}{A \code{jous} object.}

\item{...}{\dots}
}
\value{
Printed summary of the fit
}
\description{
Print a summary of \code{jous} fit.
}
